requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysql-connector-python==8.0.32
Pillow==9.4.0
protobuf==3.20.3
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
